cpu_input_boost: Improve efficiency when there is constant input
When there is constant user input while the CPU is already boosted,
instead of ignoring new input and waiting for an old boost to expire
before re-boosting, just keep an existing boost going instead of
constantly un-boosting/re-boosting the CPU. To improve efficiency when
there's constant input, only the LITTLE CPU's boost will be kept alive.
The big CPU's boost will gracefully expire in this case, and it won't have
the opportunity to be boosted again until the continuous boost finishes
and the LITTLE CPU is un-boosted.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: BijiOnta <sengolcaduk32@gmail.com>